[Texture] Reduce texture sizes tremendously are they were much bigger than necessary, using player character sprite 5984 -> 256px as reference
This makes a rescale of factor of 0.04278. We then moved sprites to children if necessary and rescaled them by the inverse factor 23.375, to restore original size as seen in game.
Traps are made of 3 sprites and many instances of traps, as well as a few instances of buttons, were not proper prefab instances so we replaced them. In the process we also simplified coloring reference: we now get Component In Children or even Components (plural) In Children for neutral platforms, so spikes children sprites can be made invisible too.